orchestrator/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.23.2
motor==3.3.2
pymongo==4.7.2
pydantic==2.7.1
esdk-obs-python==3.24.12
python-dotenv==1.0.1
requests==2.32.3
urllib3==1.26.18
dnspython==2.6.1
